Tests/LibWeb/Text/expected/async-scrolling/scroll-tree-parent-nodes.txt

scroll nodes: 2
child nodes: 1
dangling parent nodes: 0
child parent is viewport: true
